feat: Authentication to delete or update User model

+++ src/main/java/com/company/web/forummanagementsystem/service/UserServicesImpl.java
@@ -1,7 +1,9 @@
 package com.company.web.forummanagementsystem.service;
 
+import com.company.web.forummanagementsystem.exceptions.AuthorizationException;
 import com.company.web.forummanagementsystem.exceptions.DuplicateEntityException;
 import com.company.web.forummanagementsystem.exceptions.EntityNotFoundException;
+import com.company.web.forummanagementsystem.exceptions.UnauthorizedOperationException;
 import com.company.web.forummanagementsystem.models.User;
 import com.company.web.forummanagementsystem.repositories.UserRepository;
 import org.springframework.stereotype.Service;
@@ -10,6 +12,8 @@ import java.util.List;
 
 @Service
 public class UserServicesImpl implements UserServices {
+    private final static String DELETE_USER_ERROR_MESSAGE = "Only the owner of the account can delete his account!";
+    private final static String UPDATE_USER_ERROR_MESSAGE = "Only the owner of the account can change his account details!";
     private final UserRepository userRepository;
 
     public UserServicesImpl(UserRepository userRepository) {
@@ -38,13 +42,17 @@ public class UserServicesImpl implements UserServices {
     }
 
     @Override
-    public User update(User user) {
-        checkForDuplicate(user);
-        return userRepository.update(user);
+    public User update(User... users) {
+        User userToUpdate = userRepository.getById(users[0].getId());
+        checkPermissions(UPDATE_USER_ERROR_MESSAGE, userToUpdate, users[1]);
+        checkForDuplicate(users[0]);
+        return userRepository.update(users[0]);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, User user) {
+        User userToDelete = userRepository.getById(id);
+        checkPermissions(DELETE_USER_ERROR_MESSAGE, userToDelete, user);
         userRepository.delete(id);
     }
 
@@ -59,6 +67,12 @@ public class UserServicesImpl implements UserServices {
 
         if (duplicateExists) {
             throw new DuplicateEntityException("User", "email", user.getEmail());
+        }
+    }
+
+    private static void checkPermissions(String message, User... users) {
+        if (!users[0].getUsername().equals(users[1].getUsername())) {
+            throw new UnauthorizedOperationException(message);
         }
     }
 }
